use .env, only instantiate client once (venkat was right), move and fix endpoints
