+++ pages/api/connectTest.ts
@@ -5,10 +5,11 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  console.log('connecting...')
   const client = await dbConnect()
-  if (!client) res.status(500)
-  console.log('connected!')
-  const pet = await client?.db().collection('pets').findOne()
-  res.status(200).json({ pet: pet })
+  if (!client) {
+    res.status(500).json({ error: 'Client did not connect' })
+  } else {
+    res.status(200).json({ success: 'Client connected successfully' })
+  }
+  client?.close()
 }
